Improved security administration approach
